feat(solving): add bt and no strategies
Added a new BinaryCspSolver<V,D> class that synchronously solves any
binary CSP given to it using an algorithm configured by pairing a
search strategy with an ordering strategy.

Implemented the Backtracking (BT) search strategy and NO ordering
strategy for the generic binary CSP solver.

A set of 100 integration tests and 4 feature tests have been added in
this commit, in which the solving algorithm is run against binary CSPs
modelling solvable and unsolvable instances of the four problem types
(Map Colouring, N-Queens, Shikaku, Sudoku).

For a solvable puzzle, the output of modelling the puzzle as a binary
CSP then running the solver with the configured algorithm should be a
valid solution to the original puzzle.

For an unsolvable puzzle, the binary CSP solver should return a result
with no assignments, indicating a binary CSP found to have no possible
consistent assignments for all its variables.

The same framework of 100+4 tests will be used for each of the 32
solving algorithms as their constituent strategies are implemented in
forthcoming commits.