feat(solving)!: revise solver configuration
Changed the silent and verbose solving functionality so that runtime
changes to a solver's search algorithm can only be made as part of a
solving operation. This reduces the complexity of the solving type
system and reduces the risk of multi-threading race conditions.

BREAKING CHANGE: The `CheckingStrategy` and `OrderingStrategy`
properties have been removed from the silent and verbose binary CSP
solver interfaces and implementations. The `Solve` and `SolveAsync`
methods of these types have both been revised so that the client can
optionally supply a new search algorithm to replace the solver's
existing one.